# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v1.3.1
servers:
  - url: http://localhost:8080
    description: local
  - url: /histudy
    description: dev
security:
  - bearerAuth: [ ]
paths:
...
  /api/admin/edit-user:
    post:
      tags:
        - 관리자 API
      summary: 유저 정보 수정
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEdit'
        required: true
      responses:
        '200':
          description: OK
  /api/admin/form:
    delete:
      tags:
        - 관리자 API
      summary: 특정 유저 지원폼 삭제
      operationId: deleteForm
      parameters:
        - name: sid
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/admin/groupReport/{id}:
    get:
      tags:
        - 관리자 API
      summary: 특정 그룹 보고서 조회
      operationId: getTeamReports
      parameters:
        - name: id
          in: path
          description: 그룹 아이디
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamReportDto'
  /api/admin/group:
    delete:
      tags:
        - 관리자 API
      summary: 그룹 삭제
      operationId: deleteTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamIdDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
      deprecated: true
  /api/admin/academicTerm:
    post:
      tags:
        - 관리자 API
      summary: 학기 생성
      operationId: createAcademicTerm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicTermForm'
        required: true
      responses:
        '201':
          description: Created
    get:
      tags:
        - 관리자 API
      summary: 학기 목록 조회
      operationId: getAllAcademicTerms
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AcademicTermDto'
  /api/admin/academicTerm/{id}/current:
    patch:
      tags:
        - 관리자 API
      summary: 현재 학기 설정
      operationId: setCurrentTerm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Academic term ID
      responses:
        '200':
          description: OK
components:
  schemas:
...
    UserBasic:
      type: object
      properties:
        id:
          type: number
          description: User ID
          example: 1
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '211234567'
      description: List of friend added to apply form
...
    UserEdit:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          format: int64
        team:
          type: integer
          description: Group tag
          format: int32
          example: 112
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '21800012'
...
    ReportBasic:
      type: object
      properties:
        id:
          type: number
          description: Report ID
          example: 1
        title:
          type: string
          description: Report Title
          example: Week 15 Report
        regDate:
          type: string
          description: Report Last Modified Date
          example: '2021-06-01 00:00:00'
        totalMinutes:
          type: number
          description: Total minutes of the report
          example: 60
        thumbnail:
          type: string
          description: Thumbnail of the report
          example: https://histudy.s3.ap-northeast-2.amazonaws.com/2021-06-01-00-00-00-1
      description: Reports information
      example: 120
    TeamReportDto:
      type: object
      properties:
        group:
          type: number
          description: Team ID
          example: 1
          deprecated: true
        tag:
          type: integer
          description: Team tag number
          format: int32
          example: 1
        members:
          type: array
          description: Team members
          items:
            $ref: '#/components/schemas/UserBasic'
        totalTime:
          type: number
          description: Total time studied
          example: 300
        reports:
          type: array
          description: Reports information
          example: 120
          items:
            $ref: '#/components/schemas/ReportBasic'
    TeamIdDto:
      type: object
      properties:
        groupId:
          type: number
          description: Team ID
          example: 1
...
    AcademicTermForm:
      type: object
      properties:
        year:
          type: integer
          description: Academic year (YYYY)
          example: 2025
        term:
          type: string
          description: Term type (SPRING | FALL | SUMMER | WINTER)
          example: FALL
    AcademicTermDto:
      type: object
      properties:
        academicTerms:
          type: array
          items:
            $ref: '#/components/schemas/AcademicTermForm'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT