# src/main/java/edu/handong/csee/histudy/dto/AcademicTermDto.java
package edu.handong.csee.histudy.dto;

import edu.handong.csee.histudy.controller.form.AcademicTermForm;
import java.util.List;

public record AcademicTermDto(List<AcademicTermForm> academicTerms) {}